+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0-beta.5
+
+- Function and Values API
+  - Add `SassColor` class.
+
 ## 1.0.0-beta.4
 
 - Allow installing on arm64.